signature/serializers.py
from rest_framework import serializers
from .models import Document, Signature

class DocumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Document
        fields = ['id', 'title', 'file', 'uploaded_at']

class SignatureSerializer(serializers.ModelSerializer):
    class Meta:
        model = Signature
        fields = ['id', 'document', 'user', 'signed_at', 'signature_image']
